app/requirements.txt
sqlalchemy==2.0.36
pymysql==1.1.1 
pandas==2.2.3
python-dotenv==1.0.1
cryptography==44.0.0
fastapi==0.115.6
python-multipart==0.0.19
typing_extensions==4.12.2